+++ apps/client/src/app/model/user/inventory/inventory-event-type.ts
@@ -1,4 +1,5 @@
 export enum InventoryEventType {
   REMOVED,
-  ADDED
+  ADDED,
+  MOVED
 }