+++ apps/client/src/environments/patch-notes.ts
@@ -1,16 +1,12 @@
 export const patchNotes = `### Bug Fixes
 
-* **metrics:** filtering undefined source out of display.
-* **pricing:** fixed an issue with mb filling sometimes stopping.
+* **community-rotations:** fixed tags taking too much space in some browsers.
+* **desktop:** possible fix for shortcut creation, will be effective on next patch.
+* **metrics:** fixed retainer gil gains counted twice.
+* **recipe-finder:** fixed an issue with min lvl filter being readonly.
+* **trade:** changed icon priority for Allied Seals.
 
 
 ### Features
 
-* **alarms:** you can now share alarm groups.
-* **community-lists:** added tag exclusion filter.
-* **core:** support for 5.3 hotfix update.
-* **gearsets:** added scrip cost to materias table.
-* **lists:** pinned list now persists over sessions.
-* **recipe-finder:** added filters to the finder.
-* **simulator:** better indicator for unapplied consumables settings.
-* Add zone name to Gathering Items result card.`;
+* **data:** support for 5.3 hotfix price changes.`;